feat: add data in the database

+++ prisma/migrations/20230201201615_tabela_contas_criada/migration.sql
@@ -5,7 +5,7 @@ CREATE TYPE "TipoDaConta" AS ENUM ('A_Pagar', 'A_Receber');
 CREATE TABLE "Contas" (
     "id" SERIAL NOT NULL,
     "tipo" "TipoDaConta" NOT NULL,
-    "valor" INTEGER NOT NULL,
+    "valor" FLOAT NOT NULL,
     "propriedade" INTEGER NOT NULL,
     "vencimento" TIMESTAMP(3) NOT NULL,
